NEXUS-5578: When remote is scraped, prefix file is removed on discovery
This was an non handled side effect how prefix file discovery
worked, as it tried to get the prefix file (not present remotely
as we got last prefix file by scraping), and that meant that
scraping started with prefix file removed (by prefix strategy).

Now handling this case

+++ nexus-core/src/main/java/org/sonatype/nexus/proxy/maven/wl/internal/RemotePrefixFileStrategy.java
@@ -126,6 +126,17 @@ public class RemotePrefixFileStrategy
         final ResourceStoreRequest request = new ResourceStoreRequest( path );
         request.getRequestContext().put( WLManager.WL_INITIATED_FILE_OPERATION_FLAG_KEY, Boolean.TRUE );
         request.setRequestRemoteOnly( true );
+
+        // check for remote presence, as fetching with setRequestRemoteOnly has a side effect of
+        // DELETING the file from local cache if not present remotely. In this case, prefix
+        // file (on default location) obviously originates from scrape, so we should not delete it.
+        final boolean presentRemotely =
+            mavenProxyRepository.getRemoteStorage().containsItem( mavenProxyRepository, request );
+        if ( !presentRemotely )
+        {
+            return null;
+        }
+
         mavenProxyRepository.removeFromNotFoundCache( request );
         try
         {
